requirements.txt
fastmcp
mcp
dotenv
hydra-core
google-genai
pytest
pre-commit
pillow
pytest-asyncio